Fix for the EnumerationSerializer
Also add a test for concurrent deserialization of
enumerations so that we don't break it again. It
contains also a simple benchmark.

We are reflectivly finding classes and enums on each
serialization/deserialization request. It the reflection
happens to be a performance problem we can introduce
synchronised caches for the read and write path. The
introduced benchmark does not show difference though.